Extract Compiler sub right classes
